fix int conversion for latest python version

+++ atlaselectrophysiology/plot_data.py
@@ -180,7 +180,7 @@ class PlotData:
             amp_range = np.quantile(self.spikes['amps'][self.spike_idx][self.kp_idx], [0, 0.9])
             amp_bins = np.linspace(amp_range[0], amp_range[1], A_BIN)
             colour_bin = np.linspace(0.0, 1.0, A_BIN + 1)
-            colours = (cm.get_cmap('BuPu')(colour_bin)[np.newaxis, :, :3][0]) * 255
+            colours = ((cm.get_cmap('BuPu')(colour_bin)[np.newaxis, :, :3][0]) * 255).astype(np.int32)
             spikes_colours = np.empty(self.spikes['amps'][self.spike_idx][self.kp_idx].size,
                                       dtype=object)
             spikes_size = np.empty(self.spikes['amps'][self.spike_idx][self.kp_idx].size)
@@ -195,7 +195,6 @@ class PlotData:
                                     amp_bins[iA]) &
                                    (self.spikes['amps'][self.spike_idx][self.kp_idx] <=
                                     amp_bins[iA + 1]))[0]
-
                     spikes_colours[idx] = QtGui.QColor(*colours[iA])
 
                 spikes_size[idx] = iA / (A_BIN / 4)
